Update migration for apartment and fix objectid problem for physical road

+++ db/migrate/20110914160756_create_street_numbers.rb
@@ -13,6 +13,8 @@ class CreateStreetNumbers < ActiveRoad::Migration
   end
 
   def self.down
-    drop_table :street_numbers
+    if table_exists?(:street_numbers)
+      drop_table :street_numbers
+    end
   end
 end
