+++ db/migrate/20120203154500_create_junctions.rb
@@ -19,7 +19,11 @@ class CreateJunctions < ActiveRoad::Migration
   end
 
   def self.down
-    drop_table :junctions
-    drop_table :junctions_physical_roads
+    if table_exists?(:junctions)
+      drop_table :junctions
+    end
+    if table_exists?(:junctions_physical_roads)
+      drop_table :junctions_physical_roads
+    end
   end
 end
