+++ db/migrate/20121010125851_create_junction_conditionnal_costs.rb
@@ -1,6 +1,6 @@
 class CreateJunctionConditionnalCosts < ActiveRoad::Migration
   def up
-    create_table :junction_conditionnal_costs do |t|
+    create_table :junction_conditionnal_costs, :force => true do |t|
       t.belongs_to :junction
       t.float :cost
       t.string :tags
@@ -8,6 +8,8 @@ class CreateJunctionConditionnalCosts < ActiveRoad::Migration
   end
 
   def down
-    drop_table :junction_conditionnal_costs
+    if table_exists?(:junction_conditionnal_costs)
+      drop_table :junction_conditionnal_costs 
+    end
   end
 end
