+++ db/migrate/20121011124923_create_physical_road_conditionnal_costs.rb
@@ -8,6 +8,8 @@ class CreatePhysicalRoadConditionnalCosts < ActiveRecord::Migration
   end
 
   def down
-    drop_table :physical_road_conditionnal_costs
+    if table_exists?(:physical_road_conditionnal_costs)
+      drop_table :physical_road_conditionnal_costs
+    end
   end
 end
